+++ stylelint/index.mjs
@@ -130,7 +130,7 @@ export default {
 				 *
 				 * @see https://github.com/Moxio/stylelint-selector-tag-no-without-class
 				 */
-				'plugin/selector-tag-no-without-class': ['/./']
+				'plugin/selector-tag-no-without-class': [['/./'], { allowCombinators: ['>', '+', '~'] }]
 			}
 		},
 		{